% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessingTickeR.R
\name{ProcessingTickeR}
\alias{ProcessingTickeR}
\title{Processing Message Ticker}
\usage{
ProcessingTickeR(min, max, thing)
}
\arguments{
\item{min}{Starting Tick}

\item{max}{Ending Tick}

\item{thing}{Iteration Name (Null For None)}
}
\description{
Processing Message Ticker
}
