% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetPrideBar.R
\name{SetPrideBar}
\alias{SetPrideBar}
\title{Set PrideBar Starting Variables}
\usage{
SetPrideBar(min, max, flag)
}
\arguments{
\item{min}{Starting Tick}

\item{max}{Ending Tick}

\item{flag}{Pride Flag Type (See Key)}
}
\description{
Set PrideBar Starting Variables
}
